Minor notes here and there
I have decided that it is probably better to use the functionalities
offered in ml-casadi (https://github.com/TUM-AAS/ml-casadi) than my own.
It has to be somewhat appropriated to my use, but as long as that is
fine, license-wise, I prefer a tidy library to one (potentially hacky)
function. Also, my functionalities are untested, and may cause issues
that Tim Saltzmann et al. thought of, when I didn't.